Move diffusion scheduler helper to pipeline core
